requirements.txt
Flask==2.2.2
Flask-CORS==3.0.10
gunicorn==20.1.0
disnake
requests
python-dotenv
toml==0.10.2
pyyml
bs4